# chuckswapi/src/chuckswapi_api/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "chucknorris.io is a free JSON API for hand curated Chuck Norris facts.\n\nChuck Norris facts are satirical factoids about martial artist and actor Chuck Norris that have become an Internet phenomenon and as a result have become widespread in popular culture. The 'facts' are normally absurd hyperbolic claims about Norris' toughness, attitude, virility, sophistication, and masculinity.\n\nChuck Norris facts have spread around the world, leading not only to translated versions, but also spawning localized versions mentioning country-specific advertisements and other Internet phenomena. Allusions are also sometimes made to his use of roundhouse kicks to perform seemingly any task, his large amount of body hair with specific regard to his beard, and his role in the action television series Walker, Texas Ranger.",
    "version" : "2.0.0",
    "title" : "Chuck Norris IO",
    "termsOfService" : "https://api.chucknorris.io/",
    "contact" : {
      "name" : "Mathias Schilling",
      "url" : "https://www.matchilling.com",
      "email" : "m@matchilling.com"
    },
    "license" : {
      "name" : "GNU General Public License v3.0",
      "url" : "https://github.com/chucknorris-io/chuck-api/blob/master/LICENSE"
    }
  },
  "host" : "api.chucknorris.io",
  "basePath" : "/",
...
  "paths" : {
...
    "/jokes/categories" : {
      "get" : {
        "tags" : [ "joke-controller" ],
        "summary" : "getCategories",
        "operationId" : "getCategoriesUsingGET",
        "produces" : [ "application/json", "text/plain" ],
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        },
        "deprecated" : false
      }
    },
    "/jokes/random" : {
      "get" : {
        "tags" : [ "joke-controller" ],
        "summary" : "getRandomJoke",
        "operationId" : "getRandomJokeUsingGET",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "category",
          "in" : "query",
          "description" : "category",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "name",
          "in" : "query",
          "description" : "name",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/Joke"
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        },
        "deprecated" : false
      }
    },
    "/jokes/search" : {
      "get" : {
        "tags" : [ "joke-controller" ],
        "summary" : "searchValues",
        "operationId" : "searchValuesUsingGET",
        "produces" : [ "application/json", "text/plain" ],
        "parameters" : [ {
          "name" : "query",
          "in" : "query",
          "description" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "string"
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        },
        "deprecated" : false
      }
    },
    "/jokes/slack" : {
      "post" : {
        "tags" : [ "slack-controller" ],
        "summary" : "command",
        "operationId" : "commandUsingPOST",
        "consumes" : [ "application/x-www-form-urlencoded" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "channelId",
          "in" : "formData",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "channelName",
          "in" : "formData",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "command",
          "in" : "formData",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "enterpriseId",
          "in" : "formData",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "enterpriseName",
          "in" : "formData",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "responseUrl",
          "in" : "formData",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "teamDomain",
          "in" : "formData",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "teamId",
          "in" : "formData",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "text",
          "in" : "formData",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "token",
          "in" : "formData",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "triggerId",
          "in" : "formData",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "userId",
          "in" : "formData",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "userName",
          "in" : "formData",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/SlackCommandResponse"
            }
          },
          "201" : {
            "description" : "Created"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        },
        "deprecated" : false
      }
    },
    "/jokes/{id}" : {
      "get" : {
        "tags" : [ "joke-controller" ],
        "summary" : "getJokeValue",
        "operationId" : "getJokeValueUsingGET",
        "produces" : [ "text/plain" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "id",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "string"
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        },
        "deprecated" : false
      }
    }
  },
  "definitions" : {
...
    "Joke" : {
      "type" : "object",
      "required" : [ "value" ],
      "properties" : {
        "categories" : {
          "type" : "array",
          "description" : "MailingList of categories.",
          "items" : {
            "type" : "string"
          }
        },
        "created_at" : {
          "type" : "string",
          "example" : "2019-06-02 08:47:39.408742",
          "description" : "Timestamp when the joke was created."
        },
        "icon_url" : {
          "type" : "string",
          "example" : "nzf46249t8cf7wgz3rf_rg",
          "description" : "Absolute URL of the Chuck Norris icon."
        },
        "id" : {
          "type" : "string",
          "example" : "nzf46249t8cf7wgz3rf_rg",
          "description" : "URL-safe Base64-encoded UUID for a joke."
        },
        "updated_at" : {
          "type" : "string",
          "example" : "2019-06-02 08:47:39.408742",
          "description" : "Timestamp when the joke was updated."
        },
        "url" : {
          "type" : "string",
          "example" : "https://api.chucknorris.io/jokes/nzf46249t8cf7wgz3rf_rg",
          "description" : "Absolute URL of the joke."
        },
        "value" : {
          "type" : "string",
          "example" : "Chuck Norris doesn't have disk latency because the hard drive knows to hurry the hell up.",
          "description" : "The contents of an incredible funny joke."
        }
      },
      "title" : "Joke",
      "example" : {
        "icon_url" : "nzf46249t8cf7wgz3rf_rg",
        "updated_at" : "2019-06-02 08:47:39.408742",
        "created_at" : "2019-06-02 08:47:39.408742",
        "categories" : [ "categories", "categories" ],
        "id" : "nzf46249t8cf7wgz3rf_rg",
        "value" : "Chuck Norris doesn't have disk latency because the hard drive knows to hurry the hell up.",
        "url" : "https://api.chucknorris.io/jokes/nzf46249t8cf7wgz3rf_rg"
      }
    },
...
    "SlackCommandResponse" : {
      "type" : "object",
      "properties" : {
        "attachments" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/SlackCommandResponseAttachment"
          }
        },
        "icon_url" : {
          "type" : "string"
        },
        "response_type" : {
          "type" : "string"
        },
        "text" : {
          "type" : "string"
        }
      },
      "title" : "SlackCommandResponse",
      "example" : {
        "icon_url" : "icon_url",
        "attachments" : [ {
          "mrkdown_in" : [ "mrkdown_in", "mrkdown_in" ],
          "title_link" : "title_link",
          "text" : "text",
          "title" : "title",
          "fallback" : "fallback"
        }, {
          "mrkdown_in" : [ "mrkdown_in", "mrkdown_in" ],
          "title_link" : "title_link",
          "text" : "text",
          "title" : "title",
          "fallback" : "fallback"
        } ],
        "response_type" : "response_type",
        "text" : "text"
      }
    },
    "SlackCommandResponseAttachment" : {
      "type" : "object",
      "properties" : {
        "fallback" : {
          "type" : "string"
        },
        "mrkdown_in" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "text" : {
          "type" : "string"
        },
        "title" : {
          "type" : "string"
        },
        "title_link" : {
          "type" : "string"
        }
      },
      "title" : "SlackCommandResponseAttachment",
      "example" : {
        "mrkdown_in" : [ "mrkdown_in", "mrkdown_in" ],
        "title_link" : "title_link",
        "text" : "text",
        "title" : "title",
        "fallback" : "fallback"
      }
    },
...
  }
}